# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Identity Entitlement Broker API
  description: |
    Multi-tenant identity entitlement broker providing OIDC/SAML SSO integration,
    SCIM-compatible user provisioning, role mapping, entitlement management, OPA-based
    policy enforcement, and immutable audit logging.
  version: "1.0.0"
  contact:
    name: Identity Broker Team
    email: team@identitybroker.example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8081
    description: Local development server
  - url: https://api.identitybroker.example.com
    description: Production server
security:
  - bearerAuth: []
  - devAuth: []
paths:
...
  /api/v1/policy/decide:
    post:
      tags:
        - Policy
      summary: Evaluate access policy decision
      description: Sends an access request to OPA for policy evaluation. Returns allow/deny with reasoning.
      operationId: decidePolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyDecisionRequest"
            example:
              tenantId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
              actor: "jane.doe@example.com"
              subject: "jane.doe@example.com"
              action: "access"
              resource: "identity-core"
              roles:
                - "admin"
              entitlements:
                - "identity-admin"
      responses:
        "200":
          description: Policy decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyDecisionResponse"
              example:
                allowed: true
                matchedRule: "allow_access"
                denyReasons: []
                decisionTimeMs: 3
        "403":
          description: Policy denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyDecisionResponse"
              example:
                allowed: false
                matchedRule: "denied"
                denyReasons:
                  - "actor 'jane.doe@example.com' lacks entitlement for resource 'identity-core' in tenant 'a1b2c3d4-e5f6-7890-abcd-ef1234567890'"
                decisionTimeMs: 2
  /api/v1/policy/check-access:
    post:
      tags:
        - Policy
      summary: Quick access check (lightweight)
      description: Lightweight access check returning only true/false and HTTP 200/403.
      operationId: checkAccess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyDecisionRequest"
      responses:
        "200":
          description: Access granted
          content:
            application/json:
              schema:
                type: object
                properties:
                  allowed:
                    type: boolean
                example:
                  allowed: true
        "403":
          description: Access denied

  # =========================================================================
  # Audit
  # =========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from OIDC provider containing tenant_id claim
    devAuth:
      type: apiKey
      in: header
      name: X-Tenant-Id
      description: Development-only header-based tenant context (when DEV_AUTH_ENABLED=true)
  schemas:
...
    PolicyDecisionRequest:
      type: object
      required:
        - tenantId
        - actor
        - subject
        - action
        - resource
      properties:
        tenantId:
          type: string
          format: uuid
        actor:
          type: string
        subject:
          type: string
        action:
          type: string
          enum: [access, read, manage, impersonate, provision]
        resource:
          type: string
        roles:
          type: array
          items:
            type: string
        entitlements:
          type: array
          items:
            type: string
    PolicyDecisionResponse:
      type: object
      properties:
        allowed:
          type: boolean
        matchedRule:
          type: string
        denyReasons:
          type: array
          items:
            type: string
        decisionTimeMs:
          type: integer

    # -----------------------------------------------------------------------
    # Audit
    # -----------------------------------------------------------------------
...